% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhav1.R
\name{bhav1}
\alias{bhav1}
\alias{bhav}
\title{Bhavcopy from NSE upto 05 July 2024}
\source{
<https://www.nseindia.com/all-reports>
}
\usage{
bhav(x)
}
\arguments{
\item{x}{numeric date format}
}
\value{
Bhavcopy for the given date.
}
\description{
Bhavcopy from NSE upto 05 July 2024
}
\details{
Gets Bhavcopy from NSE from 1 Jan 2016 to 05 July 2024.
}
\note{
The date should be strictly numerical and mentioned in quotation mark (refer examples). `bhav1` can be used to download bhavcopy from 1 Jan 2016 to 05 July 2024.
}
\examples{
\donttest{
#Download Bhavcopy from NSE
report = bhav("01072021") # Download bhavcopy for 01 July 2021

}
}
\seealso{
\code{\link[nser]{bhav}}\code{\link[nser]{bhavtoday}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
