% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhavpr.R
\name{bhavpr}
\alias{bhavpr}
\title{Bhavcopy PR from NSE}
\source{
<https://www.nseindia.com/all-reports>
}
\usage{
bhavpr(x)
}
\arguments{
\item{x}{numeric date format}
}
\value{
List of Bhavcopy PR for the given date.
}
\description{
Bhavcopy PR from NSE
}
\details{
Gets Bhavcopy PR from NSE for the given date.
}
\note{
The date should be strictly numerical and mentioned in quotation mark, refer example.
}
\examples{
\dontrun{
# Equity Bhavcopy PR
library(nser)
report = bhavpr("01072021")
}
}
\seealso{
\code{\link[nser]{bhav}}\code{\link[nser]{bhavtoday}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
