% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhavtoday.R
\name{bhavtoday}
\alias{bhavtoday}
\title{Get Bhavcopy for the present day}
\usage{
bhavtoday()
}
\value{
Todays's Bhavcopy.
}
\description{
Get Bhavcopy for the present day
}
\details{
Gets todays bhavcopy from NSE.
}
\note{
The date should be strictly numerical and mentioned in quotation mark.The present days Bhavcopy would usually available in the evening
}
\examples{
\dontrun{
#Todays Equity Bhavcopy
library(nser)
report = bhavtoday()
}
}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhav}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
