% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailydata.R
\docType{data}
\name{dailydata}
\alias{dailydata}
\title{Daily data of a stock}
\format{
Daily data of a stock with six column
\describe{
 \item{SYMBOL}{SYMBOL of the stock}
 \item{OPEN}{Opening price for the day}
 \item{HIGH}{High price for the day}
 \item{LOW}{Low price for the day}
 \item{CLOSE}{Closing price for the day}
 \item{DATE}{Date for the given day}
}
}
\source{
https://www.nseindia.com/
}
\usage{
data(dailydata)
}
\description{
Daily data of a stock with six columns.
}
\examples{
\dontrun{
data(dailydata)
daytoweek(dailydata)
}

}
\seealso{
\code{\link{bhav}}
   ,\code{\link{bhavtoday}}
   ,\code{\link{nsetree}}
}
\keyword{datasets}
