% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytomonth.R
\name{daytomonth}
\alias{daytomonth}
\title{Convert Daily data of a stock to Monthly data}
\usage{
daytomonth(x)
}
\arguments{
\item{x}{dataframe containing stock data with columns SYMBOL, OPEN, HIGH, LOW, CLOSE & DATE.}
}
\value{
OHLC data in monthly format.
}
\description{
Convert Daily data of a stock to Monthly data
}
\details{
Converts the Daily data of a stock to Monthly data.
}
\note{
The dataframe should contain six columns named SYMBOL, OPEN, HIGH, LOW, CLOSE & DATE.
}
\examples{
\dontrun{
data(dailydata)
daytomonth(dailydata)
}

}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{bhav}}\code{\link[nser]{daytoweek}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
