% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytoweek.R
\name{daytoweek}
\alias{daytoweek}
\title{Convert Daily data of a stock to Weekly data}
\usage{
daytoweek(x)
}
\arguments{
\item{x}{dataframe containing stock data with columns SYMBOL, OPEN, HIGH, LOW, CLOSE & DATE.}
}
\value{
Dataframe with Weekly data.
}
\description{
Convert Daily data of a stock to Weekly data
}
\details{
Converts the Daily data of a stock to Weekly data.
}
\note{
The dataframe should contain six columns named SYMBOL, OPEN, HIGH, LOW, CLOSE & DATE.
}
\examples{
\donttest{
data(dailydata)
daytoweek(dailydata)
}

}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{bhav}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
