% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{Events}
\alias{Events}
\title{Upcoming events in NSE}
\usage{
Events()
}
\description{
It returns upcoming events in the national stock exchange of India.
}
\examples{
\dontrun{
df <- Events()
}
}
\author{
Deepanshu Bhalla
}
