% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdii.R
\name{fdii}
\alias{fdii}
\title{Get latest FII/DII data.}
\source{
<https://www.moneycontrol.com/stocks/marketstats/fii_dii_activity/index.php>
}
\usage{
fdii()
}
\value{
Dataframe containing FII/DII data.
}
\description{
Get latest FII/DII data.
}
\details{
Data gathered from moneycontrol.
}
\examples{
\dontrun{
# The function may fail when accessed outside from INDIA.

library(nser)
fdii()
}

}
\seealso{
\code{\link[nser]{nseipo}}\code{\link[nser]{bhav}}\code{\link[nser]{fobhavtoday}}\code{\link[nser]{nseopen}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
