% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{FilingInfo}
\alias{FilingInfo}
\title{Filing Information of financial results}
\usage{
FilingInfo(period = "quarterly")
}
\arguments{
\item{period}{"quarterly", "annual". By default it is "quarterly"}
}
\description{
It returns filing Information of financial results
}
\examples{
\dontrun{
df <- FilingInfo()
}
}
\author{
Deepanshu Bhalla
}
