% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{FinancialResults}
\alias{FinancialResults}
\title{Extract Financial Results of a Company}
\usage{
FinancialResults(Symbol)
}
\arguments{
\item{Symbol}{Symbol of the stock you are looking for. For e.g. 'RELIANCE' for Reliance Industries Limited}
}
\description{
It returns financial results of a company
}
\examples{
\dontrun{
df <- FinancialResults(Symbol = "ACC")
}
}
\author{
Deepanshu Bhalla
}
