% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fobhav.R
\name{fobhav}
\alias{fobhav}
\title{Futures and Options Bhavcopy from NSE}
\source{
<https://www.nseindia.com/all-reports>
}
\usage{
fobhav(x)
}
\arguments{
\item{x}{numeric date format}
}
\value{
F&O Bhavcopy for the given date.
}
\description{
Futures and Options Bhavcopy from NSE
}
\details{
Gets Futures and Options Bhavcopy from NSE for the given date.
}
\note{
The date should be strictly numerical and mentioned in quotation mark. `fobhav` can be used to download bhavcopy from 1 Jan 2016 onwards. To download bhavcopy previous to aforementioned date use `bhavfos`.
}
\examples{
report = fobhav("01072021") # Download F&O bhavcopy for 01 July 2021

}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
