% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fobhavtoday.R
\name{fobhavtoday}
\alias{fobhavtoday}
\title{Get F&O Bhavcopy for the present day}
\usage{
fobhavtoday()
}
\value{
Todays's F&O Bhavcopy.
}
\description{
Get F&O Bhavcopy for the present day
}
\details{
Gets todays Futures and Options bhavcopy from NSE.
}
\note{
The date should be strictly numerical and mentioned in quotation mark.The present days Bhavcopy would usually available in the evening.
}
\examples{
\dontrun{
#Todays F&O Bhavcopy
library(nser)
report = fobhavtoday()
}
}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhav}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
