% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{GetDerivativeQuote}
\alias{GetDerivativeQuote}
\title{Fetch quote of a derivate traded in NSE}
\usage{
GetDerivativeQuote(Symbol, Type = "options")
}
\arguments{
\item{Symbol}{Stock's symbol}

\item{Type}{options or future}
}
\description{
It returns information about a derivate traded in NSE.
}
\examples{
\dontrun{
QuoteDf <- GetDerivativeQuote('banknifty')
QuoteDf <- GetDerivativeQuote('banknifty', Type = "future")
}
}
\author{
Deepanshu Bhalla
}
