% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{GetQuote}
\alias{GetQuote}
\title{Fetch quote of a stock traded in NSE}
\usage{
GetQuote(Symbol)
}
\arguments{
\item{Symbol}{Stock's symbol}
}
\description{
It returns information about a stock traded in NSE.
}
\examples{
\dontrun{
QuoteDf <- GetQuote('acc')
}
}
\author{
Deepanshu Bhalla
}
