% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{Historical}
\alias{Historical}
\title{Fetch historical data of a stock traded in NSE}
\usage{
Historical(Symbol, StartDate, EndDate)
}
\arguments{
\item{Symbol}{Stock's symbol}

\item{StartDate}{Starting Date (should be in \%d-\%m-\%Y format like 20-07-2021)}

\item{EndDate}{End Date (should be in \%d-\%m-\%Y format like 24-07-2021)}
}
\description{
It returns historical information about a stock traded in NSE.
}
\examples{
\dontrun{
HistoricalDf <- Historical('ACC', "01-01-2020", "24-07-2021")
}
}
\author{
Deepanshu Bhalla
}
