% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{HistoricalDerivate}
\alias{HistoricalDerivate}
\title{Fetch historical data of derivate traded in NSE}
\usage{
HistoricalDerivate(
  Symbol,
  StartDate,
  EndDate,
  InstrumentType,
  ExpiryDate,
  StrikePrice = "",
  OptionType = ""
)
}
\arguments{
\item{Symbol}{Stock's symbol}

\item{StartDate}{Starting Date (should be in \%d-\%m-\%Y format like 20-07-2021)}

\item{EndDate}{End Date (should be in \%d-\%m-\%Y format like 24-07-2021)}

\item{InstrumentType}{"options" or "futures"}

\item{ExpiryDate}{Expiry Date (should be in \%d-\%m-\%Y format like 29-07-2021)}

\item{StrikePrice}{Strike Price}

\item{OptionType}{Option Type}
}
\description{
It returns historical information about a derivative traded in NSE.
}
\examples{
\dontrun{
HistoricalDf <- HistoricalDerivate(Symbol = 'ITC',
StartDate = "01-07-2021",
EndDate = "24-07-2021",
InstrumentType = "options",
ExpiryDate = "29-07-2021")
}
}
\author{
Deepanshu Bhalla
}
