% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{Holidays}
\alias{Holidays}
\title{Holidays List of the current year}
\usage{
Holidays(type = "trading")
}
\arguments{
\item{type}{trading or clearing}
}
\description{
It returns holidays calendar of the current year.
}
\examples{
\dontrun{
df <- Holidays()
}
}
\author{
Deepanshu Bhalla
}
