% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{Indices}
\alias{Indices}
\title{Indices in NSE}
\usage{
Indices()
}
\description{
It returns information about indices which are traded in NSE.
}
\examples{
\dontrun{
OutDf <- Indices()
}
}
\author{
Deepanshu Bhalla
}
