% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{MarketStatus}
\alias{MarketStatus}
\title{Tells if market is opened or not}
\usage{
MarketStatus()
}
\description{
It returns status if market is opened or not
}
\examples{
\dontrun{
df <- MarketStatus()
}
}
\author{
Deepanshu Bhalla
}
