% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{MostActive}
\alias{MostActive}
\title{Most Active Stocks}
\usage{
MostActive(Type = "securities", OrderBy = "volume")
}
\arguments{
\item{Type}{"securities", "etf", "sme". By default it is securities}

\item{OrderBy}{"volume", "value". By default it is volume}
}
\description{
It returns most active stocks in the national stock exchange of India.
}
\examples{
\dontrun{
df <- MostActive(OrderBy = "value")
}
}
\author{
Deepanshu Bhalla
}
