% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{NIFTY50}
\alias{NIFTY50}
\title{Stocks in Nify50}
\usage{
NIFTY50()
}
\description{
It returns information about stocks which are traded in Nifty 50.
}
\examples{
\dontrun{
OutDf <- NIFTY50()
}
}
\author{
Deepanshu Bhalla
}
