% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nseindex.R
\name{nseindex}
\alias{nseindex}
\title{NSE Nifty Indices}
\source{
<https://www1.nseindia.com/live_market/dynaContent/live_watch/live_index_watch.htm>
}
\usage{
nseindex()
}
\value{
A dataframe containing last price and % change of Nifty Indices.
}
\description{
NSE Nifty Indices
}
\details{
Live report of Nifty Indices
}
\examples{
\dontrun{
#Live status of Nifty Indices
library(nser)
nseindex()
}
}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhav}}\code{\link[nser]{fobhavtoday}}\code{\link[nser]{nseopen}}\code{\link[nser]{nselive}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
