% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsetree.R
\name{nsetree}
\alias{nsetree}
\title{Treemap for Nifty50 and F&O Securities}
\source{
<https://www.nseindia.com/market-data/live-market-indices>
}
\usage{
nsetree(x = "n50")
}
\arguments{
\item{x}{"n50" for NIFTY 50 Securities and "fo" for Futures and Options Securities.}
}
\value{
treemap A Treemap of recent percent change in value of securities.
}
\description{
Treemap for Nifty50 and F&O Securities
}
\details{
The function plots a treemap eith the recent change in price of securities.
}
\examples{
 \donttest{ # Treemap of NIFTY50 securities
#nsetree()

# Treemap of F&O securities
#nsetree("fo")
}

}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{bhavs}}\code{\link[nser]{bhavfos}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
