% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{OptionChain}
\alias{OptionChain}
\title{Option Chain}
\usage{
OptionChain(Symbol, Expiry = "all")
}
\arguments{
\item{Symbol}{Stock's symbol. You can also indices like BANKNIFTY, NIFTY etc.}

\item{Expiry}{Expiry. Try either "latest" or "all".}
}
\description{
It returns option chain.
}
\examples{
\dontrun{
}
}
\author{
Deepanshu Bhalla
}
