% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optnifty.R
\name{optnifty}
\alias{optnifty}
\title{Option Chain for NSE Nifty 50.}
\source{
<https://www.nseindia.com/option-chain>
}
\usage{
optnifty(raw_data = FALSE)
}
\arguments{
\item{raw_data}{If TRUE, the output is a raw dataframe of the option chain. Default FALSE.}
}
\value{
datatable The option chain of NIFTY 50 for the current expiry.
}
\description{
Option Chain for NSE Nifty 50.
}
\details{
Get visual friendly display of Option chain of NIFTY 50 for the current expiry.
}
\examples{
 \dontrun{# Example: NIFTY 50 option chain
optnifty()
}

}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{bhavfos}}\code{\link[nser]{nsetree}}\code{\link[nser]{optbanknifty}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
