% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse_gg.R, R/nse_options.R
\name{\%>\%}
\alias{\%>\%}
\alias{specify_decimal}
\title{nse_gg dataframe analysis}
\usage{
specify_decimal(x, k)

specify_decimal(x, k)
}
\description{
nse_gg dataframe analysis

nse options dataframe analysis
}

