% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{SecuritiesFO}
\alias{SecuritiesFO}
\title{Securities in F&O}
\usage{
SecuritiesFO()
}
\description{
It returns stocks which are traded in Future and Options.
}
\examples{
\dontrun{
OutDf <- SecuritiesFO()
}
}
\author{
Deepanshu Bhalla
}
