% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{TopGainers}
\alias{TopGainers}
\title{Top Gainers}
\usage{
TopGainers()
}
\description{
It returns list of stocks which are top gainers in the NSE.
}
\examples{
\dontrun{
df <- TopGainers()
}
}
\author{
Deepanshu Bhalla
}
