% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSEUpdated.R
\name{TopLosers}
\alias{TopLosers}
\title{Top Losers}
\usage{
TopLosers()
}
\description{
It returns top losers in the national stock exchange of India.
}
\examples{
\dontrun{
df <- TopLosers()
}
}
\author{
Deepanshu Bhalla
}
