% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhav.R
\name{bhav}
\alias{bhav}
\title{Bhavcopy from NSE}
\source{
<https://www1.nseindia.com/products/content/all_daily_reports.htm>
}
\usage{
bhav(x)
}
\arguments{
\item{x}{numeric date format}
}
\value{
Bhavcopy for the given date.
}
\description{
Bhavcopy from NSE
}
\details{
Gets Bhavcopy from NSE for the given date.
}
\note{
The date should be strictly numerical and mentioned in quotation mark. `bhav` can be used to download bhavcopy from 1 Jan 2016 onwards. To download bhavcopy previous to aforementioned date use `bhavs`.
}
\examples{
\dontrun{
report = bhav("01072021") # Download bhavcopy for 01 July 2021
head(report)
}
}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
