% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbanknifty.R
\name{optbanknifty}
\alias{optbanknifty}
\title{Option Chain for NSE BANKNIFTY.}
\source{
<https://www.nseindia.com/option-chain>
}
\usage{
optbanknifty(raw_data = FALSE)
}
\arguments{
\item{raw_data}{If TRUE, the output is a raw dataframe of the option chain. Default FALSE.}
}
\value{
datatable The option chain of BANKNIFTY for the current expiry.
}
\description{
Option Chain for NSE BANKNIFTY.
}
\details{
Get visual friendly display of Option chain of BANKNIFTY for the current expiry.
}
\examples{
 \dontrun{ # Example: BANKNIFTY option chain
optbanknifty()
}
}
\seealso{
\code{\link[nser]{bhavpr}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{bhavfos}}\code{\link[nser]{nsetree}}\code{\link[nser]{optbanknifty}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
