% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optbanknifty.R
\name{optbanknifty}
\alias{optbanknifty}
\title{Option Chain for NSE BANKNIFTY.}
\source{
<https://www.nseindia.com/option-chain>
}
\usage{
optbanknifty()
}
\value{
A dataframe containing option chain of NSE BANKNIFTY.
}
\description{
Option Chain for NSE BANKNIFTY.
}
\details{
Get visual friendly display of Option chain of BANKNIFTY for the current expiry.
}
\examples{
 \dontrun{ # Example: BANKNIFTY option chain
optbanknifty()
}
}
\seealso{
\code{\link[nser]{optnifty}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{bhavfos}}\code{\link[nser]{nsetree}}\code{\link[nser]{nseipo}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
