% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optnifty.R
\name{optnifty}
\alias{optnifty}
\title{Option Chain of NSE Nifty 50 (current expiry date).}
\source{
<https://www.nseindia.com/option-chain>
}
\usage{
optnifty()
}
\value{
A dataframe containing option chain of NSE Nifty 50 for current expiry date.
}
\description{
Option Chain of NSE Nifty 50 (current expiry date).
}
\details{
Option chain of NIFTY 50 for the current expiry.
}
\examples{
 \dontrun{# Example: NIFTY 50 option chain
optnifty()
}
}
\seealso{
\code{\link[nser]{optbanknifty}}\code{\link[nser]{bhavtoday}}\code{\link[nser]{nselive}}\code{\link[nser]{nsetree}}\code{\link[nser]{nseipo}}
}
\author{
Nandan L. Patil \email{tryanother609@gmail.com}
}
